update file name, add "ECG,PPG,GSR" to Watch page, make sure ECG and GSR are not enabled at the same time
