+++ app/src/main/java/com/mobilewizards/logging_app/MainActivity.kt
@@ -191,7 +191,7 @@ class MainActivity : AppCompatActivity() {
 
         // Set settings button
         settingsBtn.setOnClickListener{
-            val openSettings = Intent(applicationContext, SettingsNewActivity::class.java)
+            val openSettings = Intent(applicationContext, SettingsActivity::class.java)
             startActivity(openSettings)
         }
