+++ app/src/main/java/com/mobilewizards/logging_app/MyPagerAdapter.kt
@@ -6,7 +6,7 @@ import androidx.fragment.app.FragmentPagerAdapter
 
 class MyPagerAdapter(fm: FragmentManager) : FragmentPagerAdapter(fm, BEHAVIOR_RESUME_ONLY_CURRENT_FRAGMENT) {
 
-    private val fragmentList = listOf(SecondSettingPage(), FirstSettingPage())
+    private val fragmentList = listOf(PhoneSettingPage(), WatchSettingPage())
     private val fragmentTitleList = listOf("Phone", "Watch")
 
     override fun getCount(): Int = fragmentList.size
