+++ app/src/main/res/layout/activity_settings_watch.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/light_gray"
-    tools:context=".SettingsActivity">
+    tools:context=".SettingsActivityDuplicated">
 
     <ScrollView
         android:id="@+id/scroll_id"
@@ -85,35 +85,26 @@
                 android:layout_height="wrap_content"
                 android:layout_marginEnd="16dp"
                 android:text="Default"
-                app:layout_constraintBottom_toBottomOf="@+id/button_save"
+                app:layout_constraintBottom_toBottomOf="@+id/button_control"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintHorizontal_bias="1.0"
                 app:layout_constraintStart_toEndOf="@+id/settings"
-                app:layout_constraintTop_toTopOf="@+id/button_save"
+                app:layout_constraintTop_toTopOf="@+id/button_control"
                 app:layout_constraintVertical_bias="0.0" />
 
             <Button
-                android:id="@+id/button_save"
+                android:id="@+id/button_control"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="36dp"
-                android:text="Save"
+                android:text="Control"
                 app:layout_constraintBottom_toBottomOf="@+id/settings"
                 app:layout_constraintEnd_toStartOf="@+id/button_default"
                 app:layout_constraintHorizontal_bias="0.0"
                 app:layout_constraintStart_toEndOf="@+id/settings"
                 app:layout_constraintTop_toTopOf="@+id/settings" />
 
-            <Button
-                android:id="@+id/button_control"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Control"
-                app:layout_constraintTop_toBottomOf="@+id/button_default"
-                app:layout_constraintStart_toStartOf="@+id/button_default"
-                app:layout_constraintEnd_toEndOf="@+id/button_default"
-                android:layout_marginTop="5dp"
-                />
+
 
 
 
@@ -127,7 +118,7 @@
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintLeft_toLeftOf="parent"
                 app:layout_constraintRight_toRightOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/button_control"
+                app:layout_constraintTop_toBottomOf="@id/button_default"
                 android:layout_marginBottom="20dp">
 
             </LinearLayout>
